Create Board.java
This is the first draft of the Board.java file. In its current state, it can print out the user's board (the board visible to the user) and the solution board (the board which shows the ships placed). There are other methods that need to be modified because those methods are not needed in the program, such as the didIWin() method.